+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 
 * Backwards incompatible change
   * `exists?/1` helper in migration has been removed in favor of `create_if_not_exists` and `drop_if_exists`
+  * `rename table(:foo), table(:bar)` in migrations has been changed to `rename table(:foo), to: table(:bar)`
 
 ## v0.14.3